+++ src/App/Core/OpenFileHelper.cs
@@ -256,9 +256,13 @@ namespace WhereToFly.App.Core
             liveWaypointRefreshService.ClearLiveWaypointList();
             liveWaypointRefreshService.AddLiveWaypointList(locationList);
 
+            await NavigationService.Instance.NavigateAsync(Constants.PageKeyMapPage, animated: true);
+
             App.ShowToast("Locations were loaded.");
 
             App.UpdateMapLocationsList();
+
+            ZoomToLocationList(locationList);
         }
 
         /// <summary>
@@ -310,6 +314,28 @@ namespace WhereToFly.App.Core
             string direction = description.Substring(posStartDirection, posLineBreak - posStartDirection);
 
             return TakeoffDirectionsHelper.TryParse(direction, out takeoffDirections);
+        }
+
+        /// <summary>
+        /// Zooms to location list
+        /// </summary>
+        /// <param name="locationList">list of all locations</param>
+        private static void ZoomToLocationList(List<Location> locationList)
+        {
+            var minLocation = new MapPoint(90.0, 360.0, 10000.0);
+            var maxLocation = new MapPoint(-90.0, -360.0, -10000.0);
+
+            foreach (var location in locationList)
+            {
+                minLocation.Latitude = Math.Min(location.MapLocation.Latitude, minLocation.Latitude);
+                maxLocation.Latitude = Math.Max(location.MapLocation.Latitude, maxLocation.Latitude);
+                minLocation.Longitude = Math.Min(location.MapLocation.Longitude, minLocation.Longitude);
+                maxLocation.Longitude = Math.Max(location.MapLocation.Longitude, maxLocation.Longitude);
+                minLocation.Altitude = Math.Min(location.MapLocation.Altitude.Value, minLocation.Altitude.Value);
+                maxLocation.Altitude = Math.Max(location.MapLocation.Altitude.Value, maxLocation.Altitude.Value);
+            }
+
+            App.MapView.ZoomToRectangle(minLocation, maxLocation);
         }
 
         /// <summary>